Make slots for the cover to go into wider and make pole square
The pole was not a hexagon but a septagon which when the lid is flipped, the hole in it would no longer match the shape of the pole, so I made it square. I might ditch the hole in to cover and just go with a pole that remains enclosed in the box.

I also made the box wider to accomodate the wound cord, made the walls less thick as 2 mm seems like it should be enough.